Update memo_delete_confirm.php
ログイン認証追加

+++ CS1025/memo_delete_confirm.php
@@ -5,24 +5,42 @@
     // セッションを開始
     session_start();
 
-    // メモIDをGETパラメータから取得
-    $id = $_GET['id'];
+    // ログイン認証
+    if(isset($_SESSION['id'])){
 
-    // フォームが送信されたとき（削除実行時）の処理
-    if ($_SERVER["REQUEST_METHOD"] == "POST") {
-
-        //該当メモをデータベースから削除
-        $stmt = $pdo->prepare("DELETE FROM posts WHERE post_id = ?");
-        $stmt->execute([$id]);
-
-        header("Location: home.php"); // ホーム画面にリダイレクト
-        exit;
-    }
+        // メモIDをGETパラメータから取得
+        $id = $_GET['id'];
+    
+        // フォームが送信されたとき（削除実行時）の処理
+        if ($_SERVER["REQUEST_METHOD"] == "POST") {
+    
+            //該当メモをデータベースから削除
+            $stmt = $pdo->prepare("DELETE FROM posts WHERE post_id = ?");
+            $stmt->execute([$id]);
+    
+            header("Location: home.php"); // ホーム画面にリダイレクト
+            exit;
+        }
+    }else{
+        echo "ログインしていません<br />";
+        echo '<a href="../CS1025/index.php">ログイン画面へ</a>';
+    }     
 ?>
 
 <!-- 削除確認フォーム -->
-<p>このメモを削除してもよろしいですか？</p>
-<form method="post">
-    <button type="submit">OK</button>
-    <a href="memo_detail.php?id=<?= $id ?>">キャンセル</a>
-</form>
+<!DOCTYPE html>
+<html lang="ja">
+    <head>
+        <meta charset="UTF-8">
+        <title>メモ一覧</title>
+    </head>
+    <body>
+        <?php if(isset($_SESSION['id'])) : ?>
+            <p>このメモを削除してもよろしいですか？</p>
+            <form method="post">
+                <button type="submit">OK</button>
+                <a href="memo_detail.php?id=<?= $id ?>">キャンセル</a>
+            </form>
+        <?php endif ?>
+    </body>
+</html>
